fix bug product edit, authenticated

+++ controller/customer-management.js
@@ -7,16 +7,21 @@ exports.load_customer_management_page = async(req, res)=>{
 }
 
 exports.customer_delete = async(req, res)=>{
-  let id = req.params.id;
-  object_id = new ObjectId(id);
-  await customerModel.deleteOne({"_id": object_id}, function(err, res){
-    if(err){
-      console.log(err);
-    }
-    else{
-      console.log("customer is deleted");
-    }
-  });
-  res.redirect('/customer/management');
+  if(req.isAuthenticated()){
+    let id = req.params.id;
+    object_id = new ObjectId(id);
+    await customerModel.deleteOne({"_id": object_id}, function(err, res){
+      if(err){
+        console.log(err);
+      }
+      else{
+        console.log("customer is deleted");
+      }
+    });
+    res.redirect('/customer/management');
+  }
+  else{
+    res.redirect('/admin/login');
+  }
 }
