+++ controller/order-management.js
@@ -8,35 +8,51 @@ exports.load_order_management_page = async(req, res)=>{
 
 let object_id;
 exports.load_order_detail = async(req, res)=>{
-    let id = req.params.id;
-    object_id = new ObjectId(id);
-    let order = await orderModel.findOne({_id: object_id});
-    if(!order){
-        console.log("Don't find order");
+    if(req.isAuthenticated()){
+        let id = req.params.id;
+        object_id = new ObjectId(id);
+        let order = await orderModel.findOne({_id: object_id});
+        if(!order){
+            console.log("Don't find order");
+        }
+        else{
+            res.render('order-edit', {title: 'Order Detail', 'user': req.user, 'order': order});
+        }
     }
     else{
-        res.render('order-edit', {title: 'Order Detail', 'user': req.user, 'order': order});
+        res.redirect('/admin/login');
     }
 }
 
 exports.order_edit = async(req, res)=>{
-    await orderModel.updateOne({_id: object_id},{$set: {DeliveryStatus: req.body.DeliveryStatus}}, function(err){
-        if(err) {
-            console.log(err);
-            return;
+    if(req.isAuthenticated()){
+        let order = await orderModel.findOne({_id: object_id});
+        if(!order){
+            console.log("Don't find order");
         }
-    })
-    res.redirect('/order/management');
+        else{
+            await orderModel.updateOne({_id: object_id}, {$set: {DeliveryStatus: req.body.DeliveryStatus}});
+            res.redirect('/order/management');
+        }
+    }
+    else{
+        res.redirect('/admin/login');
+    }
 }
 
 exports.order_delete = async(req, res)=>{
-    let id = req.params.id;
-    let objectid = new ObjectId(id);
-    await orderModel.deleteOne({_id: objectid}, function(err){
-        if(err) {
-            console.log(err);
-            return;
-        }
-    });
-    res.redirect('/order/management');
+    if(req.isAuthenticated()){
+        let id = req.params.id;
+        let objectid = new ObjectId(id);
+        await orderModel.deleteOne({_id: objectid}, function(err){
+            if(err) {
+                console.log(err);
+                return;
+            }
+        });
+        res.redirect('/order/management');
+    }
+    else{
+        res.redirect('/admin/login');
+    }
 }
